1、完成老师成绩录入功能以及录入成绩的显示 2、完成老师登记成绩的页面功能

+++ src/main/java/com/wangjj/scoreinquirywxback/pojo/dto/UserDTO.java
@@ -27,7 +27,7 @@ public class UserDTO {
 	/** 头像 */
 	private String avatar ;
 	/**用户信息*/
-	private Integer infoId;
+	private Long infoId;
 	/** 创建人 */
 	private String createdBy ;
 	/** 创建时间 */
